+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/DemandDiscountRatingResult.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 /** DemandDiscountRatingResult */
@@ -22,7 +21,7 @@ public class DemandDiscountRatingResult {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "amount")
 
-    private BigDecimal amount;
+    private Double amount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "measure_id")
@@ -39,7 +38,7 @@ public class DemandDiscountRatingResult {
         return this;
     }
 
-    /** |参数名称：优惠标识ID| |参数约束及描述：优惠标识ID|
+    /** 优惠标识ID。
      * 
      * @return discountId */
     public String getDiscountId() {
@@ -55,7 +54,7 @@ public class DemandDiscountRatingResult {
         return this;
     }
 
-    /** |参数名称：合同商务优惠类型：基于官网价计算优惠605 华为云商务-折扣率，一口价，华为云用户606 渠道商务-折扣率，一口价，BP用户伙伴折扣优惠类型：基于官网价计算优惠607 合作伙伴授予折扣-折扣率|
+    /** 折扣优惠类型。 合同商务折扣：605：华为云BE场景下的合同商务折扣606：分销商BE场景下的合同商务折扣 伙伴授予折扣：607：合作伙伴授予折扣-折扣率
      * 
      * @return discountType */
     public Integer getDiscountType() {
@@ -66,19 +65,19 @@ public class DemandDiscountRatingResult {
         this.discountType = discountType;
     }
 
-    public DemandDiscountRatingResult withAmount(BigDecimal amount) {
+    public DemandDiscountRatingResult withAmount(Double amount) {
         this.amount = amount;
         return this;
     }
 
-    /** 优惠金额
+    /** 折扣的金额。
      * 
      * @return amount */
-    public BigDecimal getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
 
@@ -87,7 +86,7 @@ public class DemandDiscountRatingResult {
         return this;
     }
 
-    /** |参数名称：度量单位标识| |参数约束及描述：1：元|
+    /** 度量单位标识。 1：元
      * 
      * @return measureId */
     public Integer getMeasureId() {
@@ -103,7 +102,7 @@ public class DemandDiscountRatingResult {
         return this;
     }
 
-    /** |参数名称：折扣名称| |参数约束及描述：折扣名称|
+    /** 折扣名称。
      * 
      * @return discountName */
     public String getDiscountName() {
