+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/DemandProductInfo.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 /** DemandProductInfo */
@@ -57,7 +56,7 @@ public class DemandProductInfo {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "usage_value")
 
-    private BigDecimal usageValue;
+    private Double usageValue;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "usage_measure_id")
@@ -74,7 +73,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：ID标识| |参数约束及描述：同一次询价中不能重复，用于标识返回询价结果和请求的映射关系|
+    /** ID标识，同一次询价中不能重复，用于标识返回询价结果和请求的映射关系。
      * 
      * @return id */
     public String getId() {
@@ -90,7 +89,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：用户购买云服务产品的云服务类型| |参数约束及描述：例如EC2，云服务类型为hws.service.type.ec2|
+    /** 云服务类型编码，例如OBS的云服务类型编码为“hws.service.type.obs”。您可以调用查询云服务类型列表接口获取。
      * 
      * @return cloudServiceType */
     public String getCloudServiceType() {
@@ -106,8 +105,8 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：用户购买云服务产品的资源类型|
-     * |参数约束及描述：例如EC2中的VM，资源类型为hws.resource.type.vm。ResourceType是CloudServiceType中的一种资源，CloudServiceType由多种ResourceType组合提供|
+    /** 资源类型编码，例如ECS的VM为“hws.resource.type.vm”。您可以调用查询资源类型列表接口获取。
+     * ResourceType是CloudServiceType中的一种资源，CloudServiceType由多种ResourceType组合提供。
      * 
      * @return resourceType */
     public String getResourceType() {
@@ -123,7 +122,10 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：用户购买云服务产品的资源规格| |参数约束及描述：例如VM的小型规格，资源规格为m1.tiny|
+    /** 云服务类型的资源规格，部分云服务类型和资源规格举例如下：
+     * 弹性云服务器：根据操作系统类型在云服务器规格的ID后添加“.win”或“.linux”，例如“s2.small.1.linux”。云服务器规格的ID字段，您可以调用查询规格详情和规格扩展信息列表接口获取。
+     * 带宽：12_bgp：动态BGP按流量计费带宽12_sbgp：静态BGP按流量计费带宽19_bgp：动态BGP按带宽计费带宽19_sbgp：静态BGP按带宽计费带宽19_share：按带宽计费共享带宽
+     * IP：5_bgp：动态BGP公网IP5_sbgp：静态BGP公网IP 云硬盘：SATA：普通IO云硬盘SAS：高IO云硬盘GPSSD：通用型SSD云硬盘SSD：超高IO云硬盘
      * 
      * @return resourceSpec */
     public String getResourceSpec() {
@@ -139,7 +141,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：云服务区编码| |参数约束及描述：云服务区编码|
+    /** 云服务区编码，例如：“ap-southeast-1”。具体请参见地区和终端节点对应云服务的“区域”列的值。
      * 
      * @return region */
     public String getRegion() {
@@ -155,7 +157,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：可用区标识| |参数约束及描述：可用区标识|
+    /** 可用区标识。
      * 
      * @return availableZone */
     public String getAvailableZone() {
@@ -171,7 +173,8 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：资源容量大小| |参数约束及描述：例如购买的卷大小或带宽大小，只有线性产品才有这个字段| minimum: 1 maximum: 214783647
+    /** 资源容量大小，例如购买的卷大小或带宽大小。 线性产品时该参数不能为空。线性产品为包括硬盘，带宽等在订购时需要指定大小的产品。例如硬盘在订购时需选择10G、20G等不同大小。 minimum: 1 maximum:
+     * 214783647
      * 
      * @return resourceSize */
     public Integer getResourceSize() {
@@ -187,7 +190,8 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：资源容量度量标识| |参数约束及描述：枚举值如下：15：Mbps（购买带宽时使用）17：GB（购买云硬盘时使用）14：个只有线性产品才有这个字段|
+    /** 资源容量度量标识，枚举值如下： 15：Mbps（购买带宽时使用）17：GB（购买云硬盘时使用）14：个（次）
+     * 线性产品时该参数不能为空。线性产品为包括硬盘，带宽等在订购时需要指定大小的产品。例如硬盘在订购时需选择10G、20G等不同大小。
      * 
      * @return sizeMeasureId */
     public Integer getSizeMeasureId() {
@@ -203,8 +207,8 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：使用量因子编码|
-     * |参数约束及描述：云服务器：Duration云硬盘：Duration弹性IP：Duration带宽：Duration或upflow市场镜像：Duration具体每种云服务使用什么样的计费因子，需要找具体云服务确认，全集请参考|
+    /** 使用量因子编码，取值和话单中的使用量因子一致，云服务和使用量因子对应关系举例如下： 云服务器：Duration云硬盘：Duration弹性IP：Duration带宽：Duration或upflow市场镜像：Duration
+     * 您可以调用查询使用量类型列表接口获取响应参数表3中参数code的取值，即每种云服务对应的计费因子。
      * 
      * @return usageFactor */
     public String getUsageFactor() {
@@ -215,19 +219,19 @@ public class DemandProductInfo {
         this.usageFactor = usageFactor;
     }
 
-    public DemandProductInfo withUsageValue(BigDecimal usageValue) {
+    public DemandProductInfo withUsageValue(Double usageValue) {
         this.usageValue = usageValue;
         return this;
     }
 
-    /** |参数名称：使用量值| |参数约束及描述：例如按小时询价，使用量值为1，使用量单位为小时|
+    /** 使用量值。 例如按小时询价，使用量值为1，使用量单位为小时。
      * 
      * @return usageValue */
-    public BigDecimal getUsageValue() {
+    public Double getUsageValue() {
         return usageValue;
     }
 
-    public void setUsageValue(BigDecimal usageValue) {
+    public void setUsageValue(Double usageValue) {
         this.usageValue = usageValue;
     }
 
@@ -236,9 +240,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：使用量单位标识|
-     * |参数约束及描述：例如按小时询价，使用量值为1，使用量单位为小时，枚举值如下：4：小时全量枚举如下：0：天（时长）；1：元（货币）；2：角（货币）；3：分（货币）；4：小时（时长）；5：分钟（时长）；6：秒（时长）；7：EB（流量）；8：PB（流量）；9：TB（流量）；10：GB（流量）；11：MB（流量）；12：KB（流量）；13：Byte（流量）；14：个(次)（数量）；15：Mbps（流量）；16：Byte（容量）；17：GB（容量）；18：KLOC（行数）；19：年（周期）；20：月（周期）；21：MB（容量）；22：赫兹（频率）；23：核（数量）；24：天（周期）；25：小时（周期）；30：个数（个数）；31：千次（数量）；32：百万次（数量）；33：十亿次（数量）；34：bps（带宽速率）；35：kbps（带宽速率）；36：Mbps（带宽速率）；37：Gbps（带宽速率）；38：Tbps（带宽速率）；39：GB-秒（容量时长）；40：次（数量）；41：个（数量）；42：千个（数量）；43：张（数量）；44：千张（数量）；45：每秒查询率（查询速率）；46：人/天（数量）；47：TB（容量）；48：PB（容量）。具体某个云服务应该使用什么单位，需要和云服务确认。|
-     * minimum: 0 maximum: 100
+    /** 使用量度量单位，您可以调用查询度量单位列表接口获取。 例如按小时询价，使用量值为1，使用量单位为小时。 minimum: 0 maximum: 100
      * 
      * @return usageMeasureId */
     public Integer getUsageMeasureId() {
@@ -254,7 +256,7 @@ public class DemandProductInfo {
         return this;
     }
 
-    /** |参数名称：订购数量| |参数约束及描述：订购数量,有值时不能小于0，默认为1| minimum: 1 maximum: 10000
+    /** 订购数量。 minimum: 1 maximum: 10000
      * 
      * @return subscriptionNum */
     public Integer getSubscriptionNum() {
