+++ services/bssintl/src/main/java/com/huaweicloud/sdk/bssintl/v2/model/ModSubCustomerBudgetReq.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.bssintl.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 /** ModSubCustomerBudgetReq */
@@ -17,7 +16,7 @@ public class ModSubCustomerBudgetReq {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "budget_amount")
 
-    private BigDecimal budgetAmount;
+    private Double budgetAmount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "cancel_partner_frozen")
@@ -29,7 +28,7 @@ public class ModSubCustomerBudgetReq {
         return this;
     }
 
-    /** |参数名称：客户ID。| |参数约束及描述：客户ID。|
+    /** 客户账号ID。您可以调用查询客户列表接口获取customer_id。
      * 
      * @return customerId */
     public String getCustomerId() {
@@ -40,19 +39,19 @@ public class ModSubCustomerBudgetReq {
         this.customerId = customerId;
     }
 
-    public ModSubCustomerBudgetReq withBudgetAmount(BigDecimal budgetAmount) {
+    public ModSubCustomerBudgetReq withBudgetAmount(Double budgetAmount) {
         this.budgetAmount = budgetAmount;
         return this;
     }
 
-    /** |参数名称：调整的目标金额，可精确至小数点后面2位。| |参数的约束及描述：调整的目标金额，可精确至小数点后面2位。| minimum: 0 maximum: 2147483647
+    /** 调整的目标金额。 单位：元。精确至小数点后2位。 minimum: 0 maximum: 2147483647
      * 
      * @return budgetAmount */
-    public BigDecimal getBudgetAmount() {
+    public Double getBudgetAmount() {
         return budgetAmount;
     }
 
-    public void setBudgetAmount(BigDecimal budgetAmount) {
+    public void setBudgetAmount(Double budgetAmount) {
         this.budgetAmount = budgetAmount;
     }
 
@@ -61,7 +60,7 @@ public class ModSubCustomerBudgetReq {
         return this;
     }
 
-    /** |参数名称：是否在设置客户预算的同时解除账号冻结：0：否；1：是。默认为否。| |参数约束及描述：是否在设置客户预算的同时解除账号冻结：0：否；1：是。默认为否。|
+    /** 是否在设置客户预算的同时解除账号冻结： 0：否1：是 默认值为0。
      * 
      * @return cancelPartnerFrozen */
     public String getCancelPartnerFrozen() {
